python/requirements.txt
pybind11>=2.6.0
PyQt5>=5.15.0
numpy>=1.20.0
opencv-python>=4.5.0
PyOpenGL>=3.1.0
Flask>=1.1.0
Flask-Sockets>=0.2.0
glfw>=2.1.0
